Created separate Helpers file + renamed QuoteFrequencyOption -> QuoteFrequency enum
